Move transfer-related views to own module
